fix(web): feedback missing from password reset (#8021)
In the case where the password reset email takes longer than expected, a user should be given an indication as to what is happening. This disables the login form and buttons when once a user initiates the password reset and shows a loading bar. The bar is visible and the form disabled until either the email succeeds or fails. This should also help minimize duplicate emails from clicking the button multiple times.