feat(logging): reopen on sighup (#7140)
This implements a method to send Authelia a SIGHUP signal to indicate it should reload the configuration file. This can be used in two ways. If using an external logrotate service you can tell Authelia to reopen the file and create it if it doesn't exist. Secondly if you use the existing time date replacements this will create a brand new log file with the current time.

Closes #4964

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>
Signed-off-by: Brynn Crowley <littlehill723@gmail.com>
Co-authored-by: Brynn Crowley <littlehill723@gmail.com>