docs: clarify tls wording (#9077)
There is a subtle difference between the frame of reference used in client and server relationships for TLS connectivity. The technical way to refer to TLS when considering the developers frame of reference the client explicitly requests TLS i.e. StartTLS, and implicitly must use TLS in the case of a TLS socket. However this frame of reference doesn't make sense to end users. This removes the wording that could potentially be confusing in either scenario depending on the readers frame of reference.