fix(oidc): ensure stateful userinfo token use
This ensures the userinfo endpoint can be used properly when stateless introspection is enabled and the provided Access Token is a JWT Profile Access Token. Previously this caused an error which is due to the fact the session type when performing stateless introspection is arbitrary from the JWT claims.

Fixes #9382

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>