fix(configuration): default pbkdf2 iterations (#9694)
This fixes an issue where realistically the default iterations for PBKDF2 were always 310,000. But reasonably we should increase the iterations for weaker algorithms. As such the new default values are based on the NIST FIPS 140 recommendations and have been tuned to be slightly higher than those. Effectively this is a minor issue, but it deserves to be adjusted to the benefit of all users.

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>