ci(buildkite): build and grype scan on deps branches (#10657)
This change will build artifacts for renovate deps branches and include grype scanning to prevent CVEs leaking through to `master`.
There is also an update to the ignored Grype CVEs. While esbuild updates the Go version it is built with, all the detected CVEs are not actually vulnerable/relevant, but still are flagged by Grype due to stdlib and Go versioning.

Other associated changes:

* Remove dependabot configuration for github-actions
* Update renovate to pick up github-actions
* Update probot announcements to include renovate branches
* Remove conditional ignore for Buildkite steps for `renovate/` and `dependabot/` prefixed branches

Signed-off-by: Amir Zarrinkafsh <nightah@me.com>